+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.49.0] - 2023-01-31
+## Added
+- Updated types in schema
+- Added `getPinKey` and `changePINEncrypted` methods to `Card` module
+
 ## [0.48.0] - 2022-09-16
 ## Added
 - Updated types in schema